pyproject.toml
[tool.poetry]
name = "py-grpcio"
version = "0.1.0"
description = "gRPC with autogen by Pydantic models"
authors = ["Yurii <ykolibroda@lytvynov-production.com>"]
license = "MIT"
readme = "README.md"
packages = [
  { include = "py_grpcio" },
]

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "<3.0.0"
grpcio = "^1.62.2"
grpcio-tools = "^1.62.2"
protobuf = ">4.21.5,<5.0.0.dev0"
grpc-interceptor = "^0.15.4"
googleapis-common-protos = ">=1.63.0"
jinja2 = "^3.1.3"
loguru = "^0.7.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
